Add tests showing code that fails in 1.5 that worked in 1.4
It seems the type inferencing changes on arguments in 1.5 breaks a few patterns in our code.

Added two examples with `call` where the argument passed is part of a union of values, this now fails as type errors. For strings I'm able to work around it with using `as const`, but for numbers I do not think that is possible.